IE64 MMU/TLB widened to 64-bit + multi-level radix page tables
    Slice 4 of PLAN_MAX_RAM.md. Migrates the IE64 MMU surface, the
    IntuitionOS kernel's page-table layout, and the JIT block contract
    off the legacy 32-bit / flat-8192-entry table into a 6-level sparse
    radix walker over full 64-bit virtual and physical addresses. Builds
    on slice 3's uint64 bus path; MMU/TLB/PTE were intentionally untouched
    there.

    PTE format and walker (mmu_ie64.go, cpu_ie64.go):
    - PTE_PPN_SHIFT = 12, PTE_PPN_BITS = 52. Bits 0..6 are P/R/W/X/U/A/D
      flags; bits 7..11 are reserved-zero; bits 12..63 are the PPN. New
      parsePTE / makePTE helpers handle the format end-to-end.
    - 6-level sparse radix tree: top 7 bits + 5 x 9-bit interior nodes =
      52-bit VPN. Top-level table is 1 KiB at PTBR + 0..0x3FF; per-PT
      intermediate-table allocator cursor (u64) lives at
      PTBR + PT_CURSOR_OFFSET (0x400); intermediate and leaf tables grow
      from PTBR + 0x1000 upward.
    - Walker uses overflow-checked address arithmetic at every level
      (addrAddChecked, ppnToPhysChecked) so a high PTBR plus a high VPN
      cannot wrap into low memory and read an unrelated PTE; faults
      FAULT_NOT_PRESENT instead. Reads PTEs through bus.ReadPhys64 so
      intermediate tables placed in the bound Backing remain reachable.
    - translateAddr now takes uint64 vaddr and returns
      (uint64 phys, fault, cause). cpu.faultAddr widened to uint64;
      trapFault preserves a >4 GiB fault address.
    - loadMem / storeMem / execAtomic widened to uint64 vaddr. Translated
      high physes are gated by bus.PhysMapped before falling into the
      silent ReadPhys/WritePhys helpers, so unmapped data and atomic
      accesses fault rather than reading zero / writing into the void.
    - Fetch path drops IE64_ADDR_MASK; high-phys instruction fetch routes
      through bus.ReadPhys64WithFault.
    - mmuStackWrite / mmuStackRead accept uint64 sp; PUSH/POP and
      JSR/RTS/RTI work past the uint32 ceiling.

    JIT (jit_exec.go, jit_common.go):
    - Cache key is (ptbr * 0x9E3779B97F4A7C15) ^ uint64(pcVirt) so the
      full 64-bit PTBR participates and two tasks at the same PCvirt
      cannot share a compiled block.
    - needsFallback now scans every block instruction for OP_LOAD /
      STORE / FLOAD / FSTORE / DLOAD / DSTORE / JMP / JSR_IND so any
      memory-touching or dynamically targeted opcode bails the entire
      block to the interpreter. compileBlockMMU sets mmuBail on stack
      ops and direct JSR with MMU on.
    - PC-out-of-bounds path with MMU off no longer halts the CPU; it
      falls through to interpretOne so an MMU-disabled high-phys fetch
      proceeds via the bus instead of stopping the run.

    Bus / sysinfo (machine_bus.go, machine_bus_phys.go, main.go):
    - PhysMapped(addr, length) reports whether a span lies fully in low
      memory or the bound Backing. New WithFault byte/word helpers and
      PhysMapped feed the fault gates above.
    - main.go publishes ComputeMemorySizing + SetSizing +
      RegisterSysInfoMMIOFromBus right after NewMachineBus, falling back
      to DEFAULT_MEMORY_SIZE when autodetection fails so SYSINFO is
      always non-zero.
    - cpu_ie64_test.go test rig publishes the same 32 MiB sizing so
      tests' avail-style readers (SYSINFO_GUEST_RAM_PAGES) report
      non-zero pages.

    HostFS (bootstrap_hostfs.go):
    - translateGuestVA returns (uint64, bool) and walks the multi-level
      table; high PPNs are preserved verbatim instead of aliasing into
      the low 32-bit window.
    - readGuest8 / writeGuest8 gate via bus.PhysMapped before
      ReadPhys8 / WritePhys8 so a leaf PTE pointing outside both
      windows fails the HostFS call instead of returning silent zero
      or dropping the byte.

    IntuitionOS kernel (sdk/intuitionos/iexec/iexec.s,
    sdk/include/iexec.inc):
    - New helpers: kern_pt_install_leaf, kern_pt_lookup_leaf,
      kern_pt_zero_and_seed, kern_install_kern_mappings. Boot reduces
      kernel PT bring-up to two calls. Each user task PT zeroes its
      region, seeds the cursor + slot-end metadata, then replicates
      every kernel/supervisor leaf via kern_install_kern_mappings
      (full leaf replication, v0 design).
    - validate_user_range_mask, unmap_pages, safe_copy_user_name,
      every build_user_pt variant migrated to the multi-level walker.
      Permission-aware mask checks ((pte & mask) == mask) preserve
      P/R/W/X/U enforcement and the M15.6 R4 execute-only contract.
    - Per-task PT slot grew from 64 KiB to 128 KiB to hold worst-case
      multi-level intermediates + leaves (kernel mappings + USER_DYN +
      grant-installed MMIO). Introduces USER_PT_STRIDE, decoupled from
      USER_SLOT_STRIDE which still spaces user code/stack/data slots
      at 64 KiB. USER_PT_REGION_PAGES region size unchanged at 2 MiB,
      so the fixed window now holds 16 slots x 32 pages instead of 32
      slots x 16 pages; spillover into the allocator pool covers more
      tasks.
    - alloc_task_pt_block / free_task_pt_block scale automatically off
      USER_PT_SLOT_PAGES = 32; the bitmap (64 bytes / 512 bits) still
      covers the full region.
    - SYSINFO low-MMIO guest-RAM discovery wired into avail.s via
      SYSINFO_GUEST_RAM_PAGES / SYS_GET_SYS_INFO; the legacy
      MMU_NUM_PAGES dependency is gone.

    PT slot bounds-check hardening (slice 4 follow-up):
    - New PT_SLOT_END_OFFSET = 0x408 stores PTBR + slot_size_bytes
      next to the cursor. KERN_PT_STRIDE = 0x10000 (64 KiB kernel PT)
      and USER_PT_STRIDE = 0x20000 (128 KiB user PT) parameterise it.
    - kern_pt_zero_and_seed now takes R2 = slot_size_bytes and seeds
      the slot end alongside the cursor. Every install_leaf cursor
      bump bounds-checks the new cursor against the slot end and
      panics via the new pt_slot_overflow_msg banner + halt rather
      than corrupting the adjacent PT slot. All 6 zero+seed call
      sites pass the right slot size for their PT.

    Tests:
    - mmu_test_helpers_test.go installs a per-PTBR pool with cursor
      and slot-end maps, plus mmuMap / mmuLeafPTE / mmuLeafAddress /
      mmuClearLeafPTE / mmuTestSetSlotSize / mmuTestNextTableChecked
      that mirror the asm semantics for fixtures.
    - mmu_ie64_phase4a/b/c_test.go pin: PTE format round-trip, walker
      overflow guard, fault-on-unmapped data and exec, atomic high VA,
      HostFS uint64 translation, HostFS rejection of unmapped
      high-phys, JIT high-phys fetch, stack ops above 4 GiB,
      PT-slot cursor bounds (Go-side behavioural + asm source-lock).
    - iexec_test.go migrates every kernPageTableBase + page*8 reader
      to memMultiLevelLeafPTE, bumps userPT1Base by USER_PT_STRIDE,
      and adds userPtStride for synthetic two-task PT layouts.
    - Documentation: sdk/docs/IntuitionOS/IExec.md updated to describe
      the multi-level walker, the 128 KiB per-task slot, the per-PT
      cursor and slot-end metadata, and the W^X / SUA contract under
      the new format. sdk/intuitionos/CLAUDE.md security model section
      refreshed to match.

+++ m161_phase3_test.go
@@ -282,7 +282,7 @@ func resetM161Task0ClientState(client *m161Task0ClientRig) {
 	client.rig.cpu.regs[31] = client.usp
 	client.rig.cpu.userSP = client.usp
 	client.rig.cpu.kernelSP = kernStackTop
-	client.rig.cpu.ptbr = client.pt
+	client.rig.cpu.ptbr = uint64(client.pt)
 	client.rig.cpu.supervisorMode = false
 	client.rig.cpu.previousMode = false
 	client.rig.cpu.trapped = false
